Pruebas/media_analysis_app/requirements.txt
beautifulsoup4
matplotlib
numpy
pandas
playwright
python-dotenv
requests
seaborn
spacy
nltk
trafilatura
typer
gradio
wordcloud
